Address DST boundary issue
created function to get timezone offset and replaced timezoneOffsetMilliseconds variable with function call

Instead of using the timezone offset for the current date, the picker now uses the offset for the date that's currently being worked with. This prevents issues when working with a date in a different DST (see issue #2).